// frontend/projects/partner/src/app/api/model/group-assignment-dto.ts
/**
 * BDK-CoreFinance API Gateway
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Group assignment (partner-group assignment) information with assignment details
 */
export interface GroupAssignmentDto {
  /**
   * Unique group identifier - primary key
   */
  groupId: number;
  /**
   * Advisor number
   */
  advisorNumber?: string;
  /**
   * Name 1 of the group
   */
  name1?: string;
  /**
   * Name 2 of the group
   */
  name2?: string;
  /**
   * Name 3 of the group
   */
  name3?: string;
  /**
   * Participation/ownership percentage (0-100)
   */
  participation?: number;
  /**
   * KWG (Kreditwesengesetz) flag
   */
  kwgFlag?: string;
  /**
   * Engagement flag
   */
  engaFlag?: string;
  /**
   * Assignment type - required for group assignments
   */
  assignmentType?: string;
}